fix: 行列头兼容 background condition 智能反色特性
